+++ pom.xml
@@ -28,6 +28,7 @@
 
     <modules>
         <module>compiler</module>
+        <module>runtime</module>
     </modules>
 
     <properties>
