+++ runtime/pom.xml
@@ -1,0 +1,19 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>edu.clemson</groupId>
+        <artifactId>resolve-master</artifactId>
+        <version>0.0.1-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+    <artifactId>resolve-runtime</artifactId>
+    <name>RESOLVE Runtime</name>
+    <description>The RESOLVE Runtime</description>
+
+    <build>
+        <sourceDirectory>src</sourceDirectory>
+        <plugins>
+
+        </plugins>
+    </build>
+</project>
