+++ runtime/src/edu/clemson/resolve/runtime/RType.java
@@ -30,8 +30,19 @@
  */
 package edu.clemson.resolve.runtime;
 
-import java.lang.reflect.*;
+public interface RType {
 
-public interface OperationParameter extends RType {
-    public RType op(RType... e);
+    /** Returns the RESOLVE representation of a programmatic type.
+     *  @return A {@link Object} containing pointing this types representation
+     */
+    Object getRep();
+
+    void setRep(Object setRep);
+
+    /** All types in Resolve are given an initialzation value at in the modeling
+     *  phase. This initial value is implicitly applied when a variable is
+     *  declared to a Resolve {@code RType}.
+     */
+    RType initialValue();
+    // boolean representationIsPrimitive();
 }
